Fix return value of gettimeofday
The function is defined to return an int rather than a long.

+++ include/liblinux/system_calls/gettimeofday.h
@@ -3,6 +3,6 @@
 
 #include <linux/time.h>
 
-long gettimeofday(struct timeval *time_value, struct timezone *time_zone);
+int gettimeofday(struct timeval *time_value, struct timezone *time_zone);
 
 #endif /* LIBLINUX_SYSTEM_CALLS_GETTIMEOFDAY_H */
